moving files into DesktopHeader folder
